theoretical-informatics/projects/number-recognition-telegram-bot/bot_new/requirements.txt
python-telegram-bot==13.7
torch
numpy
Pillow
